refactor --> folder name = option in config
