+++ evcouplings/management/dumper/ResultsDumperInterface.py
